requirements.txt
aerich==0.7.2
aiomysql==0.2.0
aiosqlite==0.17.0
annotated-types==0.5.0
anyio==3.7.1
Authlib==1.2.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
cryptography==41.0.3
dictdiffer==0.9.0
exceptiongroup==1.1.2
fastapi==0.100.0
google-api-core==2.11.1
google-api-python-client==2.96.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.60.0
h11==0.14.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
idna==3.4
iso8601==1.1.0
itsdangerous==2.1.2
oauthlib==3.2.2
passlib==1.7.4
protobuf==4.24.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
PyJWT==2.8.0
PyMySQL==1.1.0
pyparsing==3.1.1
pypika-tortoise==0.1.6
python-multipart==0.0.6
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
setuptools==65.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tomlkit==0.11.8
tortoise-orm==0.19.3
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.23.1
